Cambios mapeo BD y en dominio para DtoReserva y entidad Reserva

+++ java-arquitectura-hexagonal/microservicio/dominio/src/main/java/com/ceiba/reserva/modelo/dto/DtoReserva.java
@@ -8,8 +8,8 @@ import java.time.LocalDateTime;
 @Getter
 @AllArgsConstructor
 public class DtoReserva {
-    private Long idReserva;
-    private Long idUsuario;
+    private int idReserva;
+    private int idUsuario;
     private double valorPagado;
     private LocalDateTime fechaUltimoPago;
 }
